Fixed some warnings, other minor fixes.

+++ source/name.cpp
@@ -157,7 +157,7 @@ void FName::StaticInit()
 	int i;
 
 	// Register hardcoded names
-	for (i = 0; i < ARRAY_COUNT(AutoNames); i++)
+	for (i = 0; i < (int)ARRAY_COUNT(AutoNames); i++)
 	{
 		Hardcode(&AutoNames[i]);
 	}
@@ -234,9 +234,12 @@ void FName::DeleteEntry(int i)
 //**************************************************************************
 //
 //	$Log$
+//	Revision 1.6  2005/04/28 07:16:15  dj_jl
+//	Fixed some warnings, other minor fixes.
+//
 //	Revision 1.5  2004/12/03 16:15:47  dj_jl
 //	Implemented support for extended ACS format scripts, functions, libraries and more.
-//
+//	
 //	Revision 1.4  2002/05/18 16:56:34  dj_jl
 //	Added FArchive and FOutputDevice classes.
 //
